Change retired `macos-13` runner to `macos-15-intel`
The `macos-13` runner was previously the way to run CI jobs on an
x86_64 (Intel) build of macOS. But it has since been deprecated and
subsequently retired. When it was deprecated, `macos-15-intel` was
introduced, which should be used instead. This makes that change to
fix CI failures such as the failure seen initially here in #21.

See https://github.com/actions/runner-images/issues/13046 for more
information on these runner images.

This commit also marks Python 3.14 non-experimental.