Change retired `macos-13` runner to `macos-15-intel` (#22)
The `macos-13` runner was previously the way to run CI jobs on an
x86_64 (Intel) build of macOS. But it has since been deprecated and
subsequently retired. When it was deprecated, `macos-15-intel` was
introduced, which should be used instead. This makes that change to
fix CI failures such as the failure seen initially here in #21.

See https://github.com/actions/runner-images/issues/13046 for more
information on these runner images.

Also fixed here:

- Don't mark Python 3.14 experimental anymore. Leave it commented
  out so it will be obvious how to add 3.15 in the nearish future.

- Fix typo that presented `amd64` instead of `arm64` as a synonym
  of `aarch64`.